restructure add_books tests and add conftest to placate flake8
