+++ openlibrary/catalog/add_book/tests/test_data/coursepuremath00hardrich_meta.mrc
